# src/main/resources/doc/openapi.json
{
  "components" : {
    "schemas" : {
...
      "edu.itmo.ultimatumgame.dto.requests.AuthenticateUserRequestDto" : {
        "properties" : {
          "id" : {
            "type" : "string"
          }
        },
        "required" : [ "id" ],
        "title" : "AuthenticateUserRequestDto",
        "type" : "object"
      },
...
      "edu.itmo.ultimatumgame.dto.requests.CreateUserRequest" : {
        "properties" : {
          "nickname" : {
            "maxLength" : 42,
            "minLength" : 3,
            "type" : "string"
          },
          "role" : {
            "default" : "PLAYER",
            "description" : "Роль пользователя. NPC на этом endpoint запрещена (403 UserRoleNotAllowedException).",
            "enum" : [ "ADMIN", "PLAYER", "OBSERVER", "NPC", "ADMIN", "PLAYER", "OBSERVER" ],
            "type" : "string"
          }
        },
        "required" : [ "nickname" ],
        "title" : "CreateUserRequest",
        "type" : "object"
      },
...
      "edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse" : {
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "error", "message", "path", "status", "timestamp" ],
        "title" : "ApiErrorResponse",
        "type" : "object"
      },
...
      "edu.itmo.ultimatumgame.dto.responses.JwtAuthenticationResponse" : {
        "properties" : {
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "token" ],
        "title" : "JwtAuthenticationResponse",
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "REST API of Ultimatum Game. Generated from source annotations.",
    "title" : "Ultimatum Game API",
    "version" : "v1"
  },
  "openapi" : "3.0.1",
  "paths" : {
    "/auth/quick-login" : {
      "post" : {
        "operationId" : "quickLogin",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.requests.AuthenticateUserRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.JwtAuthenticationResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
    "/auth/quick-register" : {
      "post" : {
        "operationId" : "quickRegister",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.requests.CreateUserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.JwtAuthenticationResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/edu.itmo.ultimatumgame.dto.responses.ApiErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}